requirements.txt
tensorflow>=2.3.0
numpy>=1.18.5
pandas>=1.1.2
matplotlib>=3.3.1
ipython>=7.18.1
six>=1.15.0
Pillow>=7.2.0
smartcrop>=0.3.2
imageio>=2.9.0
katna>=0.4.1.4
openpyxl>=3.0.5